Update deprecated usage of listener removal function
The remove all listeners functionality is deprecated
andremoved in 3.9. Therefore, the function of the
subscription has to be provided as an argument when
unsubscribing. In timeline component, a Set is added
to keep track of service, event and subscription
function, which can be used to start and stop all
listeners.