Adjust templates according to changes of role model
With the omission of the descriptor in the role model, the templates
of the user-management are adjusted to display the role's id instead.
Furthermore, in the user-data component's template, the field of the
user id is changed to type text, since the id does not necessarily need
to be a number.