Adapt user-management to new settings-service
The edit and settings route now have an object as a model,
as it is recommended. The new route no longer has a model, since the
default settings are now loaded in the user-creation component.
The settings data the user-settings component works with is now an
array of arrays. While a Map datastructure might be nicer, this data is
hard to access in a template. Objects were also not suitable, as they
are hard to type when the settingsIDs are used as properties(keys).