Add new current-user and user-settings services
The current-user service to more easily access the user. Before, the
user had to be accessed through the session service in a property
session.content.authenticated.user, which is hard to remember.
The service also provides methods for simple access of user preferences.

The new user-settings service provides a method to access an arbitrary
user's preferences and a method to access a preference's value with
the default value as fallback.