Load user and preferences in application route
Every time a user logs in or accesses the frontend with an already
valid session, the user is loaded in the current-user session.
Furthermore, to work with the user preferences, they are loaded
once from the backend. Since the user does not necessarily have
preferences, the default settings are also loaded. This allows a
one time loading of settings and prevents future queries.