Refactor user-creation to use backend's batch creation feature
To create multiple users, it is no longer necesarry to create
a record for every single user + their corresponding user preferences.
We only need one userbatchrequest record and save it.
Additionally, the function saveUser and saveMultipleUsers were
refactored to be ember-concurrency tasks, which saves the need of
the showSpinner property.