Add new user-settings-default component
This new component's only job is to display settings grouped by
their origin and add a button for every origin to turn off/on default
settings. The component for every origin displays the origin's name
the default button and passes the settings on to the
user-settings-base component, which creates the actual settings ui, i.e.
settings' names, tooltips and inputs.