Refactor user-management to load all settings dynamically
In the user management, the setting types, i.e. flagsetting/rangesetting
were hardcoded. That means, with every new type, the code would have
to be adjusted. In the application route, where settings are already
being loaded, we now push the types into a user-settings service
property. This is helpful, since we need the types to access the
settings from store. The user management components can iterate through
the types and get all settings without hardcoding them.
To support new setting types, it's only necessary to adjust the
user-settings-base template and 'tell' it how to display and mutate
the settings/data.