Fix badroute test
Visiting the badroute route in the badroute test triggers
the beforeModel function in the application route, which loads
user, settings and preferences. In the test environment, loading
the records is not possible as there is no (mocked) backend.
We now no longer load the records if in the test env.