Rework navbar button
The navbar button no longer shows a chevron when the navbar is open.
The chevron could be more easiliy confused as a back button. Thus
it's now always a hamburger icon. Furthermore, the button now has
a box shadow like the navbar itself to better indicate that they
belong together.