Fix unresolved timestamp object bug
In landscape-listener, when a new landscape arrives from the backend
via SSE, a landscape record is created and the timestamp object that
belongs to the landscape should be added to the timestamp-repo service.
However, the timestamp property of the landscape is just a PromiseObject,
which first needs to be resolved into an actual timestamp object. This
is now done and both the 'latestTimestamp' and 'timelineTimestamps'
fields of the timestamp repo service are now actually of type
Timestamp or Timestamp[], respectively.