Change and fix types in user-creation component
The 'showCreatedUsers' method had a PromiseManyArray parameter. While
there should be no issue to iterate it in this case, as the data will
not be changed by the backend, it is changed to a User Array, which is
already resolved and has the resolved users in it.
Furthermore, the class field 'roles' is now a Role Array instead of
a PromiseArray, because it is a simpler data structure to work with
and again contains resolved models.