Refactor user list to work with pagination
The list of users is passed to the user-list component via template in
the configuration.usermanagement.users template. Since the user list
has a field which keeps tracked of the checkboxes, this field needs to
be updated whenever the list of users changes (e.g. if a user is
removed). Unfortunately, there seems to be no way of really knowing
when it updates, even though the refresh method in the route does indeed
return a promise. However, this is resolved before the model in the
controller is actually updated. Therefore, we use an observer to check
for updates and then update/reset the checkboxes.
If all users on the last page are removed, we would usually be left
with an empty user list. To prevent this, we redirect the user to the
next valid page. This is done inside the controller again with an
observer.