Adapt discovery's procezz-details into native class
Additional note: The 3 actions for saving, restarting and stopping
a procezz all use the save method on the procezz and did pass
an object with a 'include' property, supposedly to tell the serializer
to include the agent relationship property. However, the save method
does not expect such property, but only a 'adapterOptions' property.
Furthermore, I could not find any difference in the sent request (url or
payload), which is why this object is no longer passed as an argument.