Extend rendering-service to govern rendering contexts and pipelines
The rendering-service now has two new Maps to maintain both
the rendering contexts (scene, renderer, camera) and rendering
pipelines, which are function arrays. Futhermore, methods were
defined for easily access data s.a. the scene. To easily apply
the the rendering pipeline, a method called 'render' was added,
which calls the functions of a pipeline sequantially.