Adjust raycaster to no longer check for type of object
Uptil now, the raycaster checked whether the object hit matched a
certain type. For an application, this could be 'component' or
'clazz'. We think this is unnecessary, since the raycasting
method already has a parameter for possible objects, it should
be made sure that this only contains the desired objects.