Outsource duplicate code in application interaction into new function
Code regarding the translation of the position to a position in the scene
is now inside an own method.
Furthermore, the methods for handling single and double click were
reduced in size and now only find the object by raycasting and
throw an event including the clicked mesh. This way, interaction is
kept simple and the event handling can be done by the component itself.