Add wip highlighting and functional closing/opening of components
A map called meshIdToModel was added to map the id of a mesh
(which represents a certain object s.a. a component) to the actual
model. That way when a component or clazz is clicked on, we
can easily access the corresponding model.
When double clicking a component, it will open or close as intended.
Instead of applying the layouting every time, we now adjust the height
and position of components/clazzes on the fly and instead of removing
the clazzes and components which are inside a closed component, we now
simply flag them invisible. This way, they are ignored during rendering.