Adjust city-layouter to always set height to closed height
The layouter before was called almost every time the user interacted
with the application. That way, the height attribute of a clazz or
component was adjusted accordingly. Since we, at best, want to call
the layouting only on initialization, the height attribute is now
the 'closed' height of the component, since the opened height is a
containt anyways. The layouter now sets the height accordingly as
the 'closed' height.