Improve timeline illustration for when there's no data
A dummy timeline is added for the period there are no actual timestamps
available. It has no data points, but the coordinate axes inlcuding
labels are, even if blurry, visible. This is ought to the enough
for the user to get an idea of what to expect while keeping
the timeline code clean free from dummy data points.
Furthermore, the message 'No timestamps available' was improved in
its illustration.