Make changes to popup components and coordinator
The popup components, like clazz-popup, were extending the popup-
coordinator and that for only one reason: inheriting the
setPopupPosition method. However, they are in no way a coordinator,
which is why this inheritance does not make much sense.
The popup components are now vanilla template-only glimmer components,
which all take the data they need as arguments and their only concern
is the illustration. The positioning of the popups is now handled by the
popup-coordinator. So its concerns are deciding which popup to display
and where to put it.
Futhermore, the positioning does no longer rely on jQuery. While it was
not per se a bad choice, it was necessary to provide a unique selector
to find the HTML Elements, in our case a class selector for both
the popover div and the parent div. This is bad for changeability, since
css class names may change and the code would then need adjustment.