Fix component/clazz rendering, add foundation mesh class
The city layouter was adjusted to return a map that contains
the resulting layouting data s.a. positions for each component
and clazz. Before, this data was directly written into the model.
However, we want to achieve a clear separation beetween data model
and visual representation. Therefor, we will use a clazz for each clazzes,
components and the foundation, which will be the THREE.JS representation,
but add more functionality s.a. highlighting and data s.a. positon data.
With this commit we first introduce the foundation-mesh as a first step.
Furthermore, the rendering of the application is working again as well as
opening and closing of components as core functionalities.

Co-authored-by: Malte Hansen <malte.hanseat@gmail.com>