Refactor users route and user-list to use DDAU pattern
Up til now, the user-list component was passed the users to be displayed
by the route and then modified them directly (i.e. deleting them).
This does, however, not match the DDAU pattern that ember uses.
The user-list now passes the action up to the routes controller, which
will do the modification. The same goes for the page-navigaion and
page-size-selector. They no longer modify the page or the size, but pass
it onto the controller.
Furthermore, we used observer functionality to listen to changes of
page, size and route model and then check if the page is in bounds and
there are actual users on that page and refresh the page accordingly.
This is now done in the afterModel hook of the users route.