Add component mesh
We introduce a specific mesh class for components which
are used in application rendering. This effort is made to
decouple view related data from the data model. Note: The
foundation is not modelled as a component mesh because a foundation
has no 'opened' state. Additionally, this allows to distinguish
components from the foundation, e.g. for raycasting.