Refactor interaction
The interaction class is now a native ES class. Instead of triggering
events from ember evented, it now, on creation, takes callback functions
for every possible event. These are optional, so only the desired
event listeners are created.
Interaction now only does the raycasting to find the mesh that is clicked
on. It no longer manipulates the state of an application and neither does
it have an instance of the popuphandler or hover effect handler.
This is up to the instantiator like application-rendering.
Keeping interaction this clean makes it easier to inherit from and
easier to reuse in general.