Use refactored interaction in application-rendering
All scene and application manipulations are now implemented
in methods insise application-rendering. These are then
passed onto the interaction instance to have them executed on
the desired events.
The Popupcoordinator is now invoked in the application-rendering
template instead of in the route. This makes it a child and we
can pass the popup content onto it with Data Down.