Apply DDAU pattern for sidebar (components)
Up til now, the additional-data service stored which components
are shown inside the sidebar. Furthermore, whether the sidebar
is shown or hidden was handled by events. To apply the DDAU pattern,
the visualization route's controller now keeps track of the
opened sidebar components and passes them down to the sidebar.
It also provides actions so that the sidebar and it's subcomponents
can close the sidebar or remove the component.
The visualization navbar with it's components like the event-opener
also get an action function to open the corresponding event-viewer.
Additional-data is now retired as it is no longer needed and was
deleted.