Do not scale line thickness in application rendering
Until now application rendering scaled the line thickness of each
communication with a factor of 0.3. This is not intuitive and it is
a better solution to modify the city layouter such that the calculated
numbers fit the need.
Additionally, some functions of the city layouter which are concerned
with communication layout are syntactically updated.