Add classes for Labels, introduce base mesh
Entity-mesh is split up in a base mesh, containing color and
highlighting informationen, and a box mesh. The box mesh is
the base class for the foundation mesh, component meshes and
clazz meshes. Additionally, a label mesh class is introduced.
This class is used as a superclass for the new clazz label mesh
and component label mesh classes.