Add application labels
Foundation labels are introduced again. Even though a
foundation is a own mesh class it uses the regular component
label meshes (since only the text color differs).
The positioning of component label is fixed and dynamically updated.
As opposed to before labels are not rotated anymore on closed
components. This is due to the fact that component meshes are
(possibly) scaled. Due to this scaling a rotated label would be
distorted. For a regular label (aligned with the parent component
mesh) the scaling can be adapted such that no distortions occur.