Remove foundation-builder and refactor foundation to not be a component
The attribute 'foundation' was removed from the Component model class.
There is no longer a foundation builder, as the foundation is
no longer created as instance of a Component.
Instead, the Application Model is now treated as if it were the
foundation, which makes sense, since the application is the top
level.