Ignore object descendents in raycasting
Up til now, the all children/descendents were including in raycasting.
This can be quite consufing, since function takes an Array with all
objects it should consider. Evenmore, the labels of the components and
clazzes are especially children of the component and clazz mesh, respectively.
To no longer consider labels inside raycasting, we removed the recursive
inclusion of all descendents of the possible objects in raycasting.