Remove base-route
Base-Route was intended to force developers creating new routes
to implement a method called resetRoute, which cleans up the
controller after the route was exited. However, this involves
the invocation of a route action, which for some unknown reason
does not work on the visualization route.
Furthermore, not all controllers need some sort of clean up, which
results in routes overriting resetRoute with an empty function.
Additionally, the resetRoute method was called twice; once on entering
and once again on leaving the route, which is unlikely intentional
and furthermore confusing.
There are two hooks inside a route: setupController and resetController,
intended for additional set up and reset of the controller. Latter can
be used just in the way as resetRoute was used.
They should be used from now on.