Fix highlighting in application-rendering
The highlighting now uses drawable clazz communication to look up
all clazzes that are involved in the communication with the selected
clazz or component. Relying on the clazzCommunications relationship
of a clazz does not work, since it only contains outgoing communications.

Furthermore, components and clazzes that are not involved in the
communication are now correctly turned transparent.