Use meshlines for application communication
Drawing application communication with comming threejs
lines is not working because the line thickness can only
be set to 1. However, we want to use different values for
the line thickness. Up to now planes with width and height
were used to imitate the desired behavior. This made the
code quite complicated and also resulted in some undesirable
drawing (e.g. not matching lines).
To get rid of this issue we use meshlines, a plugin which
internally uses triangles to draw lines.
Additionally, the code was commented and cleaned up.