Do labeling with dedicated labeling utility
Up to now the labeling for entities of the application rendering
was done by the box meshes itself. This includes the creation and
positioning of label meshes.
Since we want to reduce coupling, the task to add labels to a given
mesh is moved into a own labeler utility. This utlity provides the
necessary functions which then can be used by application-rendering.