Implement klay layouting as workers
We took the current implentation of the landscape layouting and moved
it into web workers. I.e. one worker for creating the graph for klay,
klay itself is the second worker. The third worker takes the graph
and turns the graph into data to be used by ExplorViz.
It was necessary to split the whole process into 3 workers, since
we were unable to import klay as a script inside a web worker.