Introduce custom Object3D for landscapes
To keep track of all landscape objects, every mesh which is added
to the landscape is in fact added to a landscapeObject3D. We extended
this object to a custom Object3D such that it can e.g. also remove
its own children and keep track of all added meshes.