Introduce custom Object3D for application-rendering
To keep track of all application objects, every mesh which is added
to the application is in fact added to a applicationObject3D We extended
this object to a custom Object3D such that it can e.g. also remove
its own children and keep track of all added meshes.