Make replay mode work again
The replay mode was using the application-rendering and landscape-
rendering components before refactoring and it also does now.
Both components were adjusted to have an additional mode for replay,
which has no toolbar and adds a different listener to react to an
updated replay landscape in the landscape repository service.