Fix recursive dispose method in BaseMesh
Formerly, we had a delete() method in BaseMesh, which traversed
through all ancestors of a BaseMesh and disposed all geometries and
materials. Furthermore, it deleted itself as child of its parent,
if existend. Traversing children without use of the traverse() method,
provided by inheriting from Object3D, does not work properly for
unknown reason. While the children array's length is corrent, it
is impossible to iterate over all children.
The delete() method is now a disposeRecursively() method, which, however,
does not longer removes meshes from its parents, since it breaks
the traverse() function.