Fix relative mouse position on canvas in interaction
As part of the big rendering refactoring, we wanted to access
the relative mouse position on the canvas more easily and stumbled
upon the offsetX and offsetY properties in MouseEvent's. As it
turns out, these are experimental and sometimes are zero, although
the mouse is not at the origin. To fix this, we went back to
calculate the relative position using the canvas' bounding box.