Use bootstrap-colorpicker input elements
The standard html input element for colors triggers no events while
selecting a color. An event is only triggered after a user saves the
selection (either by clicking a button or clicking outside the input
element).
This commit makes use of the bootstrap-colorpicker addon which offers
a highly customizable color input. Additionally, the
bootstrap-colorpicker offers events to register a change within the
color selection. This is used to trigger a real time update of the view.