Fix raycasting with newest threejs version
Invisible components are now considered when checking for intersections
in threejs. This causes our raycasting to return the wrong component,
i.e. components that are invisible since the parent component is
not opened. To fix this, we filter the results of the raycasting
for intersections that are visible, thus fixing the issue.