Make small adjustments to heatmap-repository service
Instead of just storing the type of the selected metric, we now store
the actual metric object inside the selectedMetric object. This way,
we can use the reference for the metric-selector component.
Furthermore, events are only triggered if the clazzMetrics could
acutally be computed, which could fail if there is no metric selected
yet.
A new class field heatmapActive is added to toggle the heatmap's legend.