Refactor application utility classes
Highlighting, entity-rendering and entity manipulation used to be
implemented as classes. However, the properties which they held were
minor ones. We decided to substitute these classes with exported functions
and either pass the needed properties or store them in the 3d application
object.
This yields the benefit that no management for the instances of the
beforementioned classes is necessary anymore.