Add button to navbar for switching between landscape- and VR-view
A new button is added to the navbar of the landscape view to enable easy switching
between the landscape- and the vr-view. The tooltip of the button gives a hint whether
WebXR is currently enabled. The button only works if that is the case.
From VR rendering one can go back to landscape rendering with a button, just as
it is implemented for application rendering.
The virtual-reality route has been removed since the seamless switching does not
require VrRendering to have a own route.