Move functionalities from vr-rendering to VRControllers
Separation of concerns is key to keep the (complex) vr-rendering component
easy to understand. Thus, we move the responsibility to raycast intersections
for controllers to the controller objects itselft. Following, the controller
objects can handle simple hovering effects and the positioning of the teleport
areas on their own.
The vr-rendering component still receives the button events of controllers
and must handle more complex behaviour which is concerned with the rendering
(like opening or closing landscape or application entities).