An abstract class called base-menu was added. It is a THREE.Mesh inhereting class, which is used to create menus in VR. For this, the menu items are drawn onto a HTML Canvas, which gets set as the Mesh's material/texture. The canvas will need to be updated to apply hover effects. The menu's job is also to find the correct item when the user interacts with the menu and call the coresponding callback function for action on the respective item.
The item class represents an item for a menu. All items need a position and
identifier (at least for now). Furthermore, it makes sure a method, which
draws its content onto a canvas, is implemented. The same goes for a method
which returns a minimum bounding rectangle, which is used for easy hover
detection.

The interactive-item class is a special item class which adds hover
functionality and the ability to add callback function, which are executed
on certain events such as hovering, pressing the trigger and holding the
trigger on  a VR controller.