Remove landscape-repository service
The service was used to store a landscape and application reference,
s.t. the rendering components knew what to display. However, this
limits us to one application-/landscape-rendering instance at a time.
Furthermore, we think it's the routes job to forward the correct
application and lanscape reference as argument to the rendering
compnents.