Delete Heatmap Rendering component
The heatmap rendering component was (for the most part) only a copy of
application-rendering. With the introduction of the in-repo addon, the
most heatmap functions are isolated in the lib directory. However,
to coordinate application and heatmap updates it is necessary and easy
to include some simple features into application-rendering which are
concerned with heatmap rendering.