Increase delay for initial resize of AR camera feed
The initial delay is increased from 100ms to 250ms. 100ms worked for most cases, however,
there were instances when the camera feed stayed unchanged with its inital resolution until
the browser window is resized. This behaviour is not tolerable for mobile devices where resizing
the browser window is no option.