Make use of raycast filter and add default filter in interaction-modifier
The passed raycast filter was not used in the modifier. This is now accounted
for. In addition, a default filter is added which removes labels and logos
from the raycasting as those are almost never of interest.
However, when passing null as a filter explicitly, the default filter is overwritten
such that all passed objects are raycasted.