Persist choice of color scheme in local storage
Until now, the changes made by a user to the color scheme are deleted
whenever the browser is reloaded. This commit saves the currently selected
color scheme to the browser's local storage.
The saved color scheme is applied on startup within the configuration service.
For this to work, the color schemes need to be accessible for the service. Therefore,
the colors are now stored within the configuration service as opposed to the colorpicker
component (where they would be inaccesible).
Individual colors are not stored as we should avoid that users really mess up
their colors, forget that they could change those, and stay behind unsatisfied.